requirements.txt
flask-mysqldb==2.0.0
python-dotenv==1.1.1